requirements.txt
streamlit
easyocr
numpy
Pillow
opencv-python-headless
torch --index-url https://download.pytorch.org/whl/cpu
torchvision --index-url https://download.pytorch.org/whl/cpu